update: model defined only in environment files
