+++ handyrl/envs/tictactoe.py
@@ -7,8 +7,82 @@ import copy
 import random
 
 import numpy as np
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
 
 from ..environment import BaseEnvironment
+from ..model import BaseModel
+
+
+class Conv(nn.Module):
+    def __init__(self, filters0, filters1, kernel_size, bn, bias=True):
+        super().__init__()
+        if bn:
+            bias = False
+        self.conv = nn.Conv2d(
+            filters0, filters1, kernel_size,
+            stride=1, padding=kernel_size//2, bias=bias
+        )
+        self.bn = nn.BatchNorm2d(filters1) if bn else None
+
+    def forward(self, x):
+        h = self.conv(x)
+        if self.bn is not None:
+            h = self.bn(h)
+        return h
+
+
+class Encoder(nn.Module):
+    def __init__(self, input_size, filters):
+        super().__init__()
+
+        self.input_size = input_size
+        self.conv = Conv(input_size[0], filters, 3, bn=False)
+
+    def forward(self, x):
+        return F.relu(self.conv(x))
+
+
+class Head(nn.Module):
+    def __init__(self, input_size, out_filters, outputs):
+        super().__init__()
+
+        self.board_size = input_size[1] * input_size[2]
+        self.out_filters = out_filters
+
+        self.conv = Conv(input_size[0], out_filters, 1, bn=False)
+        self.activation = nn.LeakyReLU(0.1)
+        self.fc = nn.Linear(self.board_size * out_filters, outputs, bias=False)
+
+    def forward(self, x):
+        h = self.activation(self.conv(x))
+        h = self.fc(h.view(-1, self.board_size * self.out_filters))
+        return h
+
+
+class SimpleConv2dModel(BaseModel):
+    def __init__(self, env, args={}):
+        super().__init__(env, args)
+
+        self.input_size = env.observation().shape
+
+        layers, filters = args.get('layers', 3), args.get('filters', 32)
+        internal_size = (filters, *self.input_size[1:])
+
+        self.encoder = Encoder(self.input_size, filters)
+        self.blocks = nn.ModuleList([Conv(filters, filters, 3, bn=True) for _ in range(layers)])
+        self.head_p = Head(internal_size, 2, self.action_length)
+        self.head_v = Head(internal_size, 1, 1)
+
+    def forward(self, x, hidden=None):
+        h = self.encoder(x)
+        for block in self.blocks:
+            h = F.relu(block(h))
+        h_p = self.head_p(h)
+        h_v = self.head_v(h)
+
+        return {'policy': h_p, 'value': torch.tanh(h_v)}
 
 
 class Environment(BaseEnvironment):
@@ -102,6 +176,9 @@ class Environment(BaseEnvironment):
 
     def players(self):
         return [0, 1]
+
+    def net(self):
+        return SimpleConv2dModel
 
     def observation(self, player=None):
         # input feature for neural nets
