+++ handyrl/train.py
@@ -22,7 +22,6 @@ import torch.optim as optim
 from .environment import prepare_env, make_env
 from .util import map_r, bimap_r, trimap_r, rotate
 from .model import to_torch, to_gpu_or_not, RandomModel
-from .model import SimpleConv2DModel as DefaultModel
 from .losses import compute_target
 from .connection import MultiProcessJobExecutor
 from .connection import accept_socket_connections
@@ -425,7 +424,7 @@ class Learner:
 
         # trained datum
         self.model_era = self.args['restart_epoch']
-        self.model_class = self.env.net() if hasattr(self.env, 'net') else DefaultModel
+        self.model_class = self.env.net()
         train_model = self.model_class(self.env, args)
         if self.model_era == 0:
             self.model = RandomModel(self.env)
